Reorganize common bits, reuse f32 for C
