+++ probe-rs/src/debug/language.rs
@@ -1,5 +1,4 @@
 use gimli::DwLang;
-use num_traits::Num;
 
 use crate::{
     debug::{DebugError, Variable, VariableCache, VariableName, VariableType, VariableValue},
@@ -10,6 +9,9 @@ use crate::{
 pub mod c;
 /// Rust
 pub mod rust;
+
+mod parsing;
+mod value;
 
 pub fn from_dwarf(dwarf_language: DwLang) -> Box<dyn ProgrammingLanguage> {
     match dwarf_language {
@@ -63,44 +65,3 @@ pub trait ProgrammingLanguage {
 pub struct UnknownLanguage;
 
 impl ProgrammingLanguage for UnknownLanguage {}
-
-/// Extension methods to simply parse a value into a number of bytes.
-trait ParseToBytes: Num {
-    type Out;
-
-    fn parse_to_bytes(s: &str) -> Result<Self::Out, DebugError>;
-}
-
-macro_rules! impl_parse {
-    ($t:ty, $bytes:expr) => {
-        impl ParseToBytes for $t
-        where
-            <$t as Num>::FromStrRadixErr: ::std::fmt::Debug,
-        {
-            type Out = [u8; $bytes];
-
-            fn parse_to_bytes(s: &str) -> Result<Self::Out, DebugError> {
-                match ::parse_int::parse::<$t>(s) {
-                    Ok(value) => Ok(<$t>::to_le_bytes(value)),
-                    Err(e) => Err(DebugError::UnwindIncompleteResults {
-                        message: format!("Invalid data conversion from value: {s:?}. {e:?}"),
-                    }),
-                }
-            }
-        }
-    };
-}
-
-impl_parse!(u8, 1);
-impl_parse!(i8, 1);
-impl_parse!(u16, 2);
-impl_parse!(i16, 2);
-impl_parse!(u32, 4);
-impl_parse!(i32, 4);
-impl_parse!(u64, 8);
-impl_parse!(i64, 8);
-impl_parse!(u128, 16);
-impl_parse!(i128, 16);
-
-impl_parse!(f32, 4);
-impl_parse!(f64, 8);
